requirements.txt
absl-py==1.3.0
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.0.1
Pillow==9.3.0
pipreqs==0.4.11
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.3.0
PyYAML==5.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post1
style==1.1.0
tensorboard==2.1.1
threadpoolctl==3.1.0
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
tqdm==4.64.1
typing_extensions==4.4.0
update==0.0.1
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.10.0
